requirements.txt
alembic==1.2.1
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.13.0
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
pycparser==2.19
pycryptodome==3.9.0
pycryptodomex==3.9.0
PyJWT==1.7.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
zipp==0.6.0
